pointerevents/pointerevent_pointermove_pointertype.html

<!doctype html>
<html>
    <head>
        <title>pointerType conservation</title>
        <meta name="viewport" content="width=device-width">
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="pointerevent_support.js"></script>
    </head>
    <body onload="run()">
        <h1>pointerType conservation</h1>
        <h4>Test Description: This test checks if pointerType attribute defined properly.</h4>
        <ol>
            <li>Press mouse left button over the black rectangle.
            <li>Then move the mouse over
        </ol>
        <p>Note: You may use several different pointer devices during the test (i.e. mouse and touch, etc.)
        <p>
        <div id="target0" style="background:black"></div>
        <script>
            var eventTested = false;
            var pointerType;
            var pointerTypeGot = false;
            var detected_pointertypes = {};
            
            setup({ explicit_done: true });
            
            var timeoutSettings = 10000;
            setup({timeout: timeoutSettings});
            
            add_completion_callback(showPointerTypes);
            
            function run() {
            	var target0 = document.getElementById("target0");
            	
            	on_event(target0, "pointerover", function(event) {
            		detected_pointertypes[ event.pointerType ] = true;
            	});
            
            	// The pointerType attribute of a pointermove event must have the same value as the pointerType attribute of the last pointerdown event with the same pointerId attribute.
            	// TA: 5.1
            	on_event(target0, "pointerdown", function (event) {
            		pointerType = event.pointerType;
            		pointerTypeGot = true;
            	});
            	
            	on_event(target0, "pointermove", function (event) {
					if(pointerTypeGot == true) {
						if(!eventTested) {
							test(function() {
									assert_true(event.pointerType == pointerType, "pointerType is dispatched properly")
								}, "pointerType is dispatched properly");
						}
						eventTested = true;
					}
            	});		
            }
        </script>
        <h1>Pointer Events pointerType conservation tests</h1>
        <div id="complete-notice">
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Would you like to <a href="javascript:location.reload()">run the tests again</a> with a different pointer type?</p>
        </div>
        <div id="log"></div>
    </body>
</html>
